### 2021-04-17 #3187 @kyleniemeyer
Hi @amandadsmith, thanks for your interest in JOSS. Unfortunately, we have a bit of a backlog, and due to a lack of suitable and available editors, I'm going to put this on our waitlist for now. We will assign this to an editor as soon as one becomes available.
