### 2021-05-25 #3187 @timtroendle
👋 @ritwikagarwal & @willu47, would any of you be willing to review this submission for JOSS?
