### 2021-05-26 #3313 @whedon
[REVIEW]: diyepw: A Python package for Do-It-Yourself EnergyPlus 1 weather file generation
**Submitting author:** @amandadsmith (<a href="http://orcid.org/0000-0003-2990-2190">Amanda D. Smith</a>)
**Repository:** <a href="https://github.com/IMMM-SFA/diyepw/" target ="_blank">https://github.com/IMMM-SFA/diyepw/</a>
**Version:** v1.2.1
