### 2021-05-26 #3313 @timtroendle
@samuelduchesne, @ritwikagarwal, we are still in reduced service mode in which we ask reviewers to finish their reviews within six weeks. I will add an automatic reminder for each of you to indicate when half of that time has elapsed.

Of course it's great if you can finish your review earlier than that.
