### 2021-06-13 #3313 @samuelduchesne
Hi @amandadsmith, just had a chance to go trough this. A tool to automatically generate weather files from NOAA data is very useful! I must say that this submission is of very good quality and it shows the care you and your team put to provide an open-source tool to the building energy modeling community.

I the spirit of making the tool even better I have some recommendations that I invite your to consider even if they would not block the acceptance of this submission.

## Continuous Integration

Although the package has automated testing, it is not part of a CI tool such as GitHub Actions, Travis CI or Circle CI. Setting up continuous integration with any of these services is very easy and will save you a lot of time (and headaches) when publishing new releases of the tool in the future. You could even test different platforms (linux, windows and mac simultaneously). My recommendation: get started right away with [Github Actions](https://docs.github.com/en/actions/guides/building-and-testing-python).

## Command Line Interface

I do not agree with issue [#31](https://github.com/IMMM-SFA/diyepw/issues/31)! Having 2 separate repositories, one for your python API and another for your CLI interface is cumbersome and some users will hate your for it. You could rather include all the cli functions as an actual shell program using [Click](https://click.palletsprojects.com/en/8.0.x/). Then users simply have to `pip install diyepw` and in the same shell, they can run `diyepw analyze_noaa_data --help` to get something similar to this:

```cmd
$ dryepw analyze_noaa_data --help
Usage: analyze_noaa_data [OPTIONS]

  Simple program that analyzes..

Options:
  --option_1 INTEGER
  --option_2 TEXT
  --option_3
```

In conclusion, I recommend the publication of this submission as is, but I strongly recommend that the authors connect the repo with a CI tool and that they merge the scripts within the same repository and implement a CLI using click.
