### 2021-06-15 IMMM-SFA/diyepw#50 @amandadsmith
Implement Continuous Integration
Reviewer suggestion: https://github.com/openjournals/joss-reviews/issues/3313#issuecomment-860224745
