### 2021-06-15 IMMM-SFA/diyepw#51 @amandadsmith
Implement cli within main repo and remove the diyepw-scripts repo
Reviewer suggestion to use Click package: https://github.com/openjournals/joss-reviews/issues/3313#issuecomment-860224745
