### 2021-06-15 #3313 @amandadsmith
Hi @samuelduchesne, we greatly appreciate your thoughtful review! These suggestions are valuable and we plan to implement. Thanks for pointing out the Click package.
