### 2021-06-15 7c950cf
add github workflow actions for running tests; fixes #50
