+++ .github/workflows/build_osx.yml
@@ -1,0 +1,34 @@
+name: osx
+
+on: [pull_request]
+
+jobs:
+  build:
+      runs-on: ${{ matrix.os }}
+
+      strategy:
+        matrix:
+          os: [macos-latest]
+
+      env:
+        OS: ${{ matrix.os }}
+        PYTHON: '3.7'
+
+      steps:
+
+        - uses: actions/checkout@v1
+
+        - name: Set up Python
+          uses: actions/setup-python@master
+          with:
+            python-version: 3.7
+
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            pip install .
+
+        - name: Test
+          run: |
+            pip install pytest
+            python -m pytest tests/
