### 2021-07-13 ce48083
merge diyepw-scripts into diyepw, using click to enable cli functionality; fixes #51
