### 2021-07-13 ba217c9
update zenodo metadata, readme badges, and github actions
