+++ .github/workflows/build_linux.yml
@@ -1,41 +1,0 @@
-name: linux
-
-on: [push]
-
-jobs:
-  build:
-      runs-on: ${{ matrix.os }}
-
-      strategy:
-        matrix:
-          os: [ubuntu-latest]
-
-      env:
-        OS: ${{ matrix.os }}
-        PYTHON: '3.7'
-
-      steps:
-
-        - uses: actions/checkout@v1
-
-        - name: Set up Python
-          uses: actions/setup-python@master
-          with:
-            python-version: 3.7
-
-        - name: Install dependencies
-          run: |
-            python -m pip install --upgrade pip
-            pip install .
-
-        - name: Test and generate coverage report
-          run: |
-            pip install pytest
-            pip install pytest-cov
-            python -m pytest tests/ --cov=./ --cov-report=xml
-
-        - name: Upload coverage to Codecov
-          uses: codecov/codecov-action@v1
-          with:
-            file: ./coverage.xml
-            fail_ci_if_error: true
