### 2021-08-13 #3313 @fneum
Hi @amandadsmith, I have just finished my review of your `diyepw` package. I very much like and support the idea to build a tool that automates the pipeline from historical weather data to uses in building energy modelling. In this case, from NOAA data to EPW files. Thanks for the extra effort you go through in making this open-source and accessible to others! I also appreciate the functionality to use heuristics to fill gaps in the data as well as the detailed tutorial.

For the record, I second both of @samuelduchesne's comments regarding continuous integration and keeping the CLI in the same repository. Great that these suggestions have both been implemented already!

I also have a few suggestions relating to a few errors I encountered while testing (PyPI packaging might be the main culprit here) and extending the introduction of the documentation (mostly using the work you put into the paper). I hope you will the review useful!

- [ ] `diyepw.get_wmo_station_location(725300)` results in an error

  ```
  Exception: Missing WMO station info file at /home/USER/miniconda3/envs/diyepw/lib/python3.9/site-packages/diyepw/data/Weather_Stations_by_County.csv
  ```

  I have used pip to install `diyepw` from PyPI. Do you ship the data folder with the `pip` installation? I think the paths in the package data might not be up to date (https://github.com/IMMM-SFA/diyepw/blob/63e8e9ad4d3a8a9490f2eb2509caa2afa7d328f6/setup.py#L25). It works if I install directly from the repository with `pip install -e .`. Could you check that?

- [ ] Is it intentional to not implement the defaults mentioned in https://diyepw.readthedocs.io/en/latest/README.html#using-diyepw-to-generate-amy-epw-files for `max_records_to_interpolate` and `max_records_to_impute` in the function `create_amy_epw_file` and similarly for `create_amy_epw_files_for_years_and_wmos`?

- [ ] When running the CLI, e.g. `create_amy_epw_files_for_years_and_wmos --help`, I get an error:

  ```
    File "/home/ws/sp2668/miniconda3/envs/diyepw/bin/create_amy_epw_files_for_years_and_wmos", line 5, in <module>
    from diyepw.scripts.create_amy_epw_files_for_years_and_wmos import create_amy_epw_files_for_years_and_wmos
  ModuleNotFoundError: No module named 'diyepw.scripts'
  ```

  Do you know what I might be doing wrong here, or whether this is a bug?

- [ ] The jupyter notebook autocomplete `shift + TAB` only shows the parameters, not the corresponding explanations. It would be nice to have them show up there. Would that be simple to fix? (minor)

- [ ] Any reason why `analyze_noaa_isd_lite_files` has compulsory arguments, but `analyze_noaa_isd_lite_file` does not? I would have expected both to function widely analogously. (minor)

- [ ] The versions in `setup.py` are mostly fixed. Any chance you can loosen these constraints? (minor)

- [ ] The first paragraph of the README/documentation directly uses the abbreviations EPW and WMO. For the latter, could you mention what the abbreviation stands for, for the former, could you link to an explanation of the file format and add some background where these files are commonly used?

- [ ] In general, the introduction is quite minimal. Sure, it doesn't need to be long, but given the documentation/README will be the primary point of entry for prospective users, I think it would be helpful to copy over some of the background, context, statement of need from the paper.

- [ ] Particularly, the data source NOAA ISD Lite could be featured more prominently.

- [ ] Could you link to a website that lists the available codes for the WMOs? Actually, it would be useful to have such a reference also in the docstring of functions like `create_amy_epw_files_for_years_and_wmos`. I found a link in the tutorial, but currently it's somewhat hidden.

- [ ] Please also add a page like "Contributing" with community guidelines (one of the boxes I need to tick above)

- [ ] You could link to the documentation in the **About** section of the Github repository (top right) (minor)

- [ ] L21: "Its output is *a* weather file"

- [ ] L25-26: The separation of `diyepw` and `diyepw-scripts` seems to be outdated: https://github.com/IMMM-SFA/diyepw/issues/51

- [ ] L51f: Could you add another sentence on the functionality/background of LAF (apart from the limitations)?

- [ ] references: many links are duplicated, could you remove the duplicates?

- [ ] references: could you round out the citations of `numpy`, `pandas`, `xarray`? I.e. with DOI following instructions at https://numpy.org/citing-numpy/, https://pandas.pydata.org/about/citing.html, http://xarray.pydata.org/en/stable/getting-started-guide/faq.html#how-should-i-cite-xarray
