### 2021-08-22 #3313 @timtroendle
@amandadsmith, I will be out of office for the duration mentioned above. Do you think you will be able to handle @fneum's comments before that?
