### 2021-08-23 IMMM-SFA/diyepw#55 @amandadsmith
JOSS review #2: Documentation
This is from the JOSS review thread: openjournals/joss-reviews#3313
Functionality issues reproduced below for convenience with comments from @amandadsmith

> ## Documentation
> * [x]  The first paragraph of the README/documentation directly uses the abbreviations EPW and WMO. For the latter, could you mention what the abbreviation stands for, for the former, could you link to an explanation of the file format and add some background where these files are commonly used?

- EnergyPlus Weather
- World Meteorological Organization
File format defined: https://bigladdersoftware.com/epx/docs/8-3/auxiliary-programs/energyplus-weather-file-epw-data-dictionary.html
Note: Can use material from the paper itself for explanation in docs

> * [x]  In general, the introduction is quite minimal. Sure, it doesn't need to be long, but given the documentation/README will be the primary point of entry for prospective users, I think it would be helpful to copy over some of the background, context, statement of need from the paper.
Note: Can use material from the paper itself for explanation in docs. Reach out to @amandadsmith for content as needed

> * [x]  Particularly, the data source NOAA ISD Lite could be featured more prominently.
> * [x]  Could you link to a website that lists the available codes for the WMOs? Actually, it would be useful to have such a reference also in the docstring of functions like `create_amy_epw_files_for_years_and_wmos`. I found a link in the tutorial, but currently it's somewhat hidden.

I am not aware of a website that has a current, dependable list. @cdburley referred me to a journal paper for this. We may need to check with him about whether we should qualify this link in some way if we're not confident it lists all the available codes.

> * [x]  Please also add a page like "Contributing" with community guidelines (one of the boxes I need to tick above)
> * [x]  You could link to the documentation in the **About** section of the Github repository (top right) (minor)
>
