### 2021-08-23 IMMM-SFA/diyepw#55 @BenjaminStuermer
Regarding the master list of WMO codes, it looks like maybe they have it here? http://www.weathergraphics.com/identifiers/
