### 2021-08-23 bb8e186
Correct numpy and xarray citations
Per https://numpy.org/citing-numpy/ and http://xarray.pydata.org/en/stable/getting-started-guide/faq.html#how-should-i-cite-xarray
