+++ paper/paper.bib
@@ -21,7 +21,6 @@
   booktitle    = "Pacific Northwest National Laboratory",
   year         =  2021,
   url          = "https://im3.pnnl.gov/",
-  howpublished = "\url{https://im3.pnnl.gov/}",
   note         = "Accessed: 2021-3-24"
 }
 
@@ -69,7 +68,6 @@
                   Office (BTO)}",
   year         =  2020,
   url          = "https://energyplus.net/",
-  howpublished = "\url{https://energyplus.net/}",
   note         = "Accessed: 2021-3-16"
 }
 
@@ -80,7 +78,6 @@
   publisher    = "\{Big Ladder Software LLC\}",
   year         =  2015,
   url          = "https://bigladdersoftware.com/epx/docs/8-3/auxiliary-programs/energyplus-weather-file-epw-data-dictionary.html",
-  howpublished = "\url{https://bigladdersoftware.com/epx/docs/8-3/auxiliary-programs/energyplus-weather-file-epw-data-dictionary.html}",
   note         = "Accessed: 2021-3-24"
 }
 
@@ -90,7 +87,6 @@
   publisher    = "National Oceanic and Atmospheric Administration",
   year         =  2021,
   url          = "https://www.ncdc.noaa.gov/isd",
-  howpublished = "\url{https://www.ncdc.noaa.gov/isd}",
   note         = "Accessed: 2021-3-24"
 }
  
@@ -107,7 +103,6 @@
   title        = "Weather Data",
   booktitle    = "{EnergyPlus}",
   url          = "https://energyplus.net/weather",
-  howpublished = "\url{https://energyplus.net/weather}",
   note         = "Accessed: 2021-3-24"
 }
 
@@ -116,7 +111,6 @@
   title        = "Weather Data for Simulation",
   booktitle    = "{EnergyPlus}",
   url          = "https://energyplus.net/weather/simulation",
-  howpublished = "\url{https://energyplus.net/weather/simulation}",
   note         = "Accessed: 2021-3-24"
 }
 
@@ -145,7 +139,6 @@
   month        =  apr,
   year         =  2021,
   url          = "https://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt",
-  howpublished = "\url{https://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt}"
 }
 
 
