+++ paper/paper.bib
@@ -107,10 +107,12 @@
 }
 
 
-@MISC{eplus-weather-data-for-simulation,
+@MISC{EPwds,
   title        = "Weather Data for Simulation",
   booktitle    = "{EnergyPlus}",
+  author       = "{U.S. Department of Energy}",
   url          = "https://energyplus.net/weather/simulation",
+  howpublished = "\url{https://energyplus.net/weather/simulation}",
   note         = "Accessed: 2021-3-24"
 }
 
