+++ paper/paper.md
@@ -56,7 +56,7 @@ observed weather data to drive simulations of model buildings using EnergyPlus f
 - LAF relies on an API for downloading observed weather data that has limitations on the amount of data that can be downloaded without a paid account.
 - LAF is not directly extensible to other sources of weather data, such as the NOAA ISD Lite format used here.
 
-The EnergyPlus website lists additional resources for obtaining EPW files for building energy modeling [@eplus-weather-data-for-simulation]. Few data providers can produce weather files for specific locations over a given historical period, and when they do provide such EPW files, the raw data representing the weather observations may not be available to the user. Thus the processing of that data to produce the EPWs is not fully transparent and reproducible. The user may be required to pay for these files and would not have the option to adjust the standards for data quality--for determining which values are acceptable for a given meteorological variable, or for limiting the amount of data that is interpolated or otherwise imputed by the software generating the EPW files.
+The EnergyPlus website lists additional resources for obtaining EPW files for building energy modeling [@EPwds]. Few data providers can produce weather files for specific locations over a given historical period, and when they do provide such EPW files, the raw data representing the weather observations may not be available to the user. Thus the processing of that data to produce the EPWs is not fully transparent and reproducible. The user may be required to pay for these files and would not have the option to adjust the standards for data quality--for determining which values are acceptable for a given meteorological variable, or for limiting the amount of data that is interpolated or otherwise imputed by the software generating the EPW files.
 
 # Dependencies
 
