+++ paper/paper.bib
@@ -99,12 +99,15 @@
 }
 
 
-@MISC{eplus-weather-data,
+
+@MISC{EPweather,
   title        = "Weather Data",
   booktitle    = "{EnergyPlus}",
+  author       = "{U.S. Department of Energy}",
   url          = "https://energyplus.net/weather",
-  note         = "Accessed: 2021-3-24"
+  note         = "Accessed: 2021-8-23"
 }
+
 
 
 @MISC{EPwds,
