### 2021-08-23 d622f79
Remove outdated info
separation of diyepw and diyepw-scripts no longer the case
