### 2021-08-23 19f198e
Remove github_diyepw-scripts reference
no longer referenced in paper
