+++ paper/paper.bib
@@ -7,14 +7,6 @@
 }
 
 
-@MISC{github_diyepw-scripts,
-  title       = "diyepw-scripts",
-  publisher   = "Github",
-  institution = "Pacific Northwest National Laboratory",
-  year        =  2021,
-  url         = "https://github.com/IMMM-SFA/diyepw-scripts"
-}
-
 
 @MISC{IM3web,
   title        = "Integrated Multisector Multiscale Modeling",
