### 2021-08-23 8050f94
add mesowest/mesonet citations
