+++ paper/paper.bib
@@ -181,3 +181,22 @@
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
+
+
+@MISC{mesowest,
+  title        = "{{MesoWest} Data}",
+  author       = "{University of Utah Department of Atmospheric Sciences}",
+  publisher    = "University of Utah",
+  year         =  2021,
+  url          = "http://mesowest.utah.edu/",
+  note         = "Accessed: 2021-8-23"
+}
+
+
+@MISC{mesonet,
+  title        = "Mesonet {API}",
+  author       = "{Synoptic Data PBC}",
+  url          = "https://developers.synopticdata.com/mesonet/",
+  note         = "Accessed: 2021-8-23"
+}
+
