### 2021-08-24 a559407
Include description of what LAF is
