+++ paper/paper.bib
@@ -151,6 +151,46 @@
   url       = {http://doi.org/10.5334/jors.148}
 }
 
+@software{xarray-zenodo,
+  author       = {Stephan Hoyer and
+                  Joe Hamman and
+                  Maximilian Roos and
+                  Deepak Cherian and
+                  Clark Fitzgerald and
+                  keewis and
+                  Keisuke Fujii and
+                  Fabien Maussion and
+                  crusaderky and
+                  Alex Kleeman and
+                  Spencer Clark and
+                  Thomas Kluyver and
+                  Mathias Hauser and
+                  James Munroe and
+                  Tom Nicholas and
+                  Zac Hatfield-Dodds and
+                  Ryan Abernathey and
+                  MaximilianR and
+                  Phillip J. Wolfram and
+                  alexamici and
+                  Julia Signell and
+                  Yohai Bar Sinai and
+                  Jonathan J. Helmus and
+                  Kai Mühlbauer and
+                  Markel and
+                  Gerardo Rivera and
+                  Pete Cable and
+                  Tom Augspurger and
+                  johnomotani and
+                  Benoit Bovy},
+  title        = {pydata/xarray: v0.16.2},
+  month        = nov,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.16.2},
+  doi          = {10.5281/zenodo.4299126},
+  url          = {https://doi.org/10.5281/zenodo.4299126}
+}
+
 @manual{pandas,
   title = {{{pandas}}: Flexible and powerful data analysis / manipulation library for Python},
   author = {{pandas contributors}},
