### 2021-08-24 1ef024a
Correct pandas citation, add pandas-zenodo
per https://pandas.pydata.org/about/citing.html
