+++ paper/paper.bib
@@ -191,13 +191,29 @@
   url          = {https://doi.org/10.5281/zenodo.4299126}
 }
 
-@manual{pandas,
-  title = {{{pandas}}: Flexible and powerful data analysis / manipulation library for Python},
-  author = {{pandas contributors}},
-  year = {2021},
-  organization = {{PyData}},
-  type = {Manual}
+@InProceedings{pandas,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
 }
+
+
+@software{pandas-zenodo,
+    author       = {The pandas development team},
+    title        = {pandas-dev/pandas: Pandas},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {latest},
+    doi          = {10.5281/zenodo.3509134},
+    url          = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+
 
 @Article{         numpy,
  title         = {Array programming with {NumPy}},
