### 2021-08-24 4a13f68
Update pandas citations
per https://pandas.pydata.org/about/citing.html
