+++ paper/paper.bib
@@ -203,7 +203,7 @@
 
 
 @software{pandas-zenodo,
-    author       = {The pandas development team},
+    author       = "{The pandas development team}",
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
