+++ paper/paper.bib
@@ -204,15 +204,14 @@
 
 @software{pandas-zenodo,
     author       = "{The pandas development team}",
-    title        = {pandas-dev/pandas: Pandas},
-    month        = feb,
-    year         = 2020,
-    publisher    = {Zenodo},
-    version      = {latest},
-    doi          = {10.5281/zenodo.3509134},
-    url          = {https://doi.org/10.5281/zenodo.3509134}
+  title        = {pandas-dev/pandas: Pandas 1.3.2},
+  month        = aug,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v1.3.2},
+  doi          = {10.5281/zenodo.5203279},
+  url          = {https://doi.org/10.5281/zenodo.5203279}
 }
-
 
 
 @Article{         numpy,
