### 2021-08-24 af55c3c
update documentation based on joss review; closes #55
