### 2021-08-24 IMMM-SFA/diyepw#55 @amandadsmith
@thurber For the [readme](https://diyepw.readthedocs.io/en/feature-review/README.html): Can we get the HTML to serve all the URLs as live links? e.g. in markdown
```
- [Browse Weather Data](https://energyplus.net/weather)
- [Weather Station Identifiers](http://www.weathergraphics.com/identifiers/)
```
