+++ setup.py
@@ -22,7 +22,7 @@ setuptools.setup(  # pragma: no cover
       author='Amanda Smith',
       author_email='amanda.d.smith@pnnl.gov',
       packages=setuptools.find_packages(),
-      package_data={'diyepw': ['data/**', 'test/files/**']},
+      package_data={'diyepw': ['data/**', 'test/files/**', 'data/**/*', 'test/files/**/*']},
       license='BSD 2-Clause',
       python_requires='~=3.7',
       install_requires=[
