### 2021-08-25 #3313 @thurber
Hi @fneum, thanks for reporting these issues! We've released `diyepw` v1.2.1 to PyPi that should address these points:
* `diyepw.get_wmo_station_location(725300)` should be fixed - yes, this was issue in the setup.py with pathfinding.
* Implemented the defaults in most of the methods.
* Hopefully fixed the CLI issues -- if you're still seeing this can you let me know what OS you're running on?
* Jupyter autocomplete -- we were missing the :param: documentation in some of the methods, so we've added this in
* For `analyze_noaa_isd_lite_file(s)` methods, we added the default parameters to the plural version, but decided to still leave them out of the singular version in case consuming workflows want to make their own decisions about that.
* Updated documentation based on your feedback (added information and links about EPW, WMO, and ISD, as well as a CONTRIBUTING.md section)

Regarding fixed dependency versions in `setup.py` -- this is understandably a pain point, but we find it important to pin the versions in order to favor full reproducibility for any experiments making use of the code. `diyepw` will most likely still work with alternate dependency versions, but you'll need to manually install in that case. We recommend maintaining a unique virtual environment for `diyepw` workflows.
