### 2021-08-25 #3313 @amandadsmith
HI @fneum we really appreciate your support and your suggestions here. I'll skip the Functionality items as I believe @thurber has touched on all of them above. I believe we've gotten each of the Documentation/Paper issues, and I think it's more clear and up to date, so thanks again for your thoroughness!

>
> ## Documentation
> * [x ]  The first paragraph of the README/documentation directly uses the abbreviations EPW and WMO. For the latter, could you mention what the abbreviation stands for, for the former, could you link to an explanation of the file format and add some background where these files are commonly used?
> * [ x]  In general, the introduction is quite minimal. Sure, it doesn't need to be long, but given the documentation/README will be the primary point of entry for prospective users, I think it would be helpful to copy over some of the background, context, statement of need from the paper.

We've added a little material here but will also link directly from this page to the paper when published.

> * [x ]  Particularly, the data source NOAA ISD Lite could be featured more prominently.
> * [x ]  Could you link to a website that lists the available codes for the WMOs? Actually, it would be useful to have such a reference also in the docstring of functions like `create_amy_epw_files_for_years_and_wmos`. I found a link in the tutorial, but currently it's somewhat hidden.
> * [x ]  Please also add a page like "Contributing" with community guidelines (one of the boxes I need to tick above)
> * [ x]  You could link to the documentation in the **About** section of the Github repository (top right) (minor)
>
> ## Paper
> * [x]  L21: "Its output is _a_ weather file"
> * [x ]  L25-26: The separation of `diyepw` and `diyepw-scripts` seems to be outdated: [Implement cli within main repo and remove the diyepw-scripts repo IMMM-SFA/diyepw#51](https://github.com/IMMM-SFA/diyepw/issues/51)
> * [x ]  L51f: Could you add another sentence on the functionality/background of LAF (apart from the limitations)?
> * [x ]  references: many links are duplicated, could you remove the duplicates?
> * [ ]  references: could you round out the citations of `numpy`, `pandas`, `xarray`? I.e. with DOI following instructions at https://numpy.org/citing-numpy/, https://pandas.pydata.org/about/citing.html, http://xarray.pydata.org/en/stable/getting-started-guide/faq.html#how-should-i-cite-xarray
