### 2021-08-28 #3313 @thurber
@timtroendle, here's the DOI for `v1.2.1`: [10.5281/zenodo.5258122](https://doi.org/10.5281/zenodo.5258122)
