### 2021-08-28 #3313 @amandadsmith
> Thank you @fneum for your review.
>
> @amandadsmith, it all looks good to me. Is `v1.2.1` the correct version? At this point, could you please
>
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
https://github.com/openjournals/joss-reviews/issues/3313#issuecomment-907672011
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
I confirm the title matches, author list is correct and includes ORCIDs.
> * [x]  Please list the DOI of the archived version here.
https://github.com/openjournals/joss-reviews/issues/3313#issuecomment-907672011
>
> I can then move forward with accepting the submission.
Thank you!
