### 2021-08-31 #3313 @kyleniemeyer
Hi @amandadsmith, I'm the AEIC on duty this week, doing some final checks before accepting. I noticed that the EnergyPlus reference has some extra curly brackets around the title—can you remove those? Everything else looks good!
