+++ paper/paper.bib
@@ -67,7 +67,7 @@
 @MISC{EPWdd,
   title        = "{EnergyPlus} Weather File ({EPW}) Data Dictionary",
   booktitle    = "Auxiliary Programs --- {EnergyPlus} 8.3",
-  publisher    = "\{Big Ladder Software LLC\}",
+  publisher    = "Big Ladder Software LLC",
   year         =  2015,
   url          = "https://bigladdersoftware.com/epx/docs/8-3/auxiliary-programs/energyplus-weather-file-epw-data-dictionary.html",
   note         = "Accessed: 2021-3-24"
