### 2021-08-31 #3313 @amandadsmith
Hi @kyleniemeyer, these extra curly brackets have been removed. Thank you!
